+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.4.10] - 2025-03-10
+
+### Added
+- Add `getOrInsert()` and `getOrInsertComputed()` for `Map` and `WeakMap`
+- Add `TaskController` and `TaskSignal` polyfills
+
+### Changed
+- Update `scheduler.yield()` to be compatible with the spec
+
+### Deprecated
+- Warn that `Map.prototype.emplace()` was removed from the proposal
+
 ## [v0.4.9] - 2025-01-22
 
 ### Added